client-mac/requirements.txt
pywebview==4.4.1
requests==2.32.3
